Major changes
* Moved everything to "./include/thread_pool"

* Main include file is now:
  <thread_pool/thread_pool.hpp>

* Everything is now in the `tp` namespace

* Added optional `boost::future` compatiblity with
  `THREAD_POOL_USE_BOOST` macro

* Added `make install` command to CMakeFiles.txt

* Adapted tests and benchmarks to new changes
