+++ include/thread_pool.hpp
@@ -1,0 +1,3 @@
+#pragma once
+
+#include <thread_pool/thread_pool.hpp>
